update after refactor4

+++ database/models.py
@@ -23,7 +23,7 @@ class StreamEmails(Base):
 class User(Base):
     __tablename__ = "users"
 
-    id = Column(Integer, primary_key=True, index=True, autoincrement=True, unique=True)
+    id = Column(BigInteger, primary_key=True, index=True, autoincrement=True, unique=True)
     telegram_id: int = Column(BigInteger, nullable=False, unique=True)
     telegram_username = Column(String(255), nullable=True, unique=True)
     balance_amount = Column(Float, nullable=False, default=0)
